requirements.txt
git+https://github.com/KurimuzonAkuma/pyrogram.git
TgCrypto==1.2.5 
motor==3.1.2
dnspython==2.3.0
hachoir==3.2.0
Pillow==9.5.0 
aiohttp==3.8.4
pytz==2023.3
asyncio 
hachoir
ffmpeg-python
